Been working on implementing double-RX/TX. NOT fully functioning.
Still need to fix following:
Manager queue instead of multiprocessing queue to be able to handle two processes reading from the same queue. This queue does not have a close method. 
Can not concatenate bytes to a str in fragment when this type of packet is sent. Look into this. 
The old processes don't seem to be killable.