Attempt at tracking out how much bandwidth we are achieving.
Tried using nrf.writeFast(data). This proved useless as it only read the first.. 300? or so bytes. It gives a high TX as it does not care about the queue size or buffer at all, so it simply writes as fast as the little computer on the Pi is able to. Will need to keep working on proving speed.